backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
moviepy
gTTS
openai
python-multipart
Pillow